+++ SDLut/src/SDLJoystick.cc
@@ -79,7 +79,7 @@ std::string JoystickPool::getName(int index)
 	std::map<Joystick::JoyHat,short>::iterator itEnd = JoyHatrage2sdlmap.end();
 	for (; it != itEnd; ++it)
 	{
-		assert((*it).first >= 0 && (*it).first < result.size());
+		assert((*it).first >= 0 && (*it).first < static_cast<int>(result.size()));
 		result[(*it).first] = (*it).second;
 	}
 }
