+++ SDLut/src/SDLMouse.cc
@@ -30,7 +30,7 @@ namespace RAGE
 	std::map<Mouse::Button, short>::iterator itEnd = Buttonrage2sdlmap.end();
 	for (; it != itEnd; ++it)
 	{
-		assert((*it).first >= 0 && (*it).first < result.size());
+		assert((*it).first >= 0 && (*it).first < static_cast<int>(result.size()));
 		result[(*it).first] = (*it).second;
 	}
 }
